requirements.txt
# Core dependencies
numpy>=1.25.0
pandas>=2.1.1
scikit-learn>=1.4.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Deep Learning
torch>=2.3.0
torchvision>=0.18.0

# Reinforcement Learning
ray[rllib]>=2.9.0
gymnasium>=0.29.1

# API and Web
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.5.0

# Cybersecurity tools
pyshark>=0.6.0
scapy>=2.5.0
elasticsearch>=8.10.0
elasticsearch-dsl>=8.9.0

# MLOps
mlflow>=2.9.0
wandb>=0.16.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
joblib>=1.3.0

# Dashboarding
dash>=2.8.0
dash-bootstrap-components>=1.4.0
plotly>=5.13.0

# Database
sqlalchemy>=2.0.0

# System Monitoring
psutil>=5.9.0